Rename kernel headers
